Revert "Create copies."
This reverts commit 2bf76eaa03657a82234bd8e75ae3fc2cf840ea81.
